feat: add query search hook as draft
