+++ src/inputs/dropdown/search-dropdown/PSearchDropdown.stories.mdx
@@ -377,7 +377,7 @@ export const Template = (args, { argTypes }) => ({
 ## Playground
 
 <Canvas>
-    <Story name="playground" args={{menu: getSearchDropdownMenu()}}>
+    <Story name="playground" args={{menu: getSearchDropdownMenu()}} height={'300px'}>
         {Template.bind({})}
     </Story>
 </Canvas>
